Fixed bug with counting repeats
I'm using regexp to count sequences without overlaps quickly and this
requires padding the sequences with spaces.  This wasn't done properly
so in some repetitive sequences I was getting the wrong count.  This
should be fixed now with an explanation in the comments.  I've also
added a test to compressSequenceNFast to pick up this kind of error if
it crops up again.  The test just makes sure that the size of the
compressed sequence and the calculated savings is consistent with the
size of the original sequence and the discovered grammar.